Restrutura funções de pip para diretório dedicado
